refactor(feeling thunk tests): taking feeling thunk tests out of user thunk test file into its own spec
