+++ client/store/user.spec.js
@@ -93,44 +93,3 @@ describe('Opinion Thunk Creators', () => {
     })
   })
 })
-
-describe('Feeling Thunk Creators', () => {
-  let store
-  let mockAxios
-
-  const initialState = []
-
-  beforeEach(() => {
-    mockAxios = new MockAdapter(axios)
-    store = mockStore(initialState)
-  })
-
-  afterEach(() => {
-    mockAxios.restore()
-    store.clearActions()
-  })
-
-  describe('getFeelings', () => {
-    it('eventually dispatches the GET_FEELINGS action', () => {
-      const fakeFeeling = [{name: 'Love', category: 'verb'}]
-      mockAxios.onGet('/api/feelings').replyOnce(200, fakeFeeling)
-      return store.dispatch(getFeelings()).then(() => {
-        const actions = store.getActions()
-        expect(actions[0].type).to.be.equal('GET_FEELINGS')
-        expect(actions[0].feelings).to.be.deep.equal(fakeFeeling)
-      })
-    })
-  })
-
-  describe('addFeeling', () => {
-    it('eventually dispatches the ADD_FEELING action', () => {
-      const newFeeling = {name: 'Hate', category: 'verb'}
-      mockAxios.onPost('/api/feelings').replyOnce(204, newFeeling)
-      return store.dispatch(createFeeling(newFeeling)).then(() => {
-        const actions = store.getActions()
-        expect(actions[0].type).to.be.equal('ADD_FEELING')
-        expect(actions[0].feeling).to.be.deep.equal(newFeeling)
-      })
-    })
-  })
-})
